src/requirements.txt
attrs==19.3.0
bokeh==1.4.0
branca==0.4.0
Cartopy==0.17.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
dash==1.10.0
dash-core-components==1.9.0
dash-html-components==1.0.3
dash-renderer==1.3.0
dash-table==4.6.2
Django==2.2.12
django-leaflet==0.26.0
ephem==3.7.7.1
Fiona==1.8.13.post1
Flask==1.1.2
Flask-Compress==1.4.0
folium==0.10.1
future==0.18.2
geopandas==0.7.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
munch==2.5.0
numpy==1.18.2
packaging==20.3
pandas==0.24.2
Pillow==7.1.1
plotly==4.6.0
psycopg2==2.8.5
pyparsing==2.4.7
pyproj==2.6.0
pyshp==2.1.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
retrying==1.3.3
scipy==1.4.1
Shapely==1.7.0
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.8
Werkzeug==1.0.1
